Rename coverage options, remove TestContext (unsound), move files around
