+++ packages/xstate-test/test/index.test.ts
@@ -1,562 +1,7 @@
 import { createTestModel } from '../src';
 import { assign, createMachine, interpret } from 'xstate';
 import { getDescription } from '../src/utils';
-import { stateValueCoverage, transitionCoverage } from '../src/coverage';
-
-interface DieHardContext {
-  three: number;
-  five: number;
-}
-
-const pour3to5 = assign<DieHardContext>((ctx) => {
-  const poured = Math.min(5 - ctx.five, ctx.three);
-
-  return {
-    three: ctx.three - poured,
-    five: ctx.five + poured
-  };
-});
-const pour5to3 = assign<DieHardContext>((ctx) => {
-  const poured = Math.min(3 - ctx.three, ctx.five);
-
-  const res = {
-    three: ctx.three + poured,
-    five: ctx.five - poured
-  };
-
-  return res;
-});
-const fill3 = assign<DieHardContext>({ three: 3 });
-const fill5 = assign<DieHardContext>({ five: 5 });
-const empty3 = assign<DieHardContext>({ three: 0 });
-const empty5 = assign<DieHardContext>({ five: 0 });
-
-const dieHardMachine = createMachine<DieHardContext>(
-  {
-    id: 'dieHard',
-    initial: 'pending',
-    context: { three: 0, five: 0 },
-    states: {
-      pending: {
-        always: {
-          target: 'success',
-          cond: 'weHave4Gallons'
-        },
-        on: {
-          POUR_3_TO_5: {
-            actions: pour3to5
-          },
-          POUR_5_TO_3: {
-            actions: pour5to3
-          },
-          FILL_3: {
-            actions: fill3
-          },
-          FILL_5: {
-            actions: fill5
-          },
-          EMPTY_3: {
-            actions: empty3
-          },
-          EMPTY_5: {
-            actions: empty5
-          }
-        },
-        meta: {
-          description: (state) => {
-            return `pending with (${state.context.three}, ${state.context.five})`;
-          },
-          test: async ({ jugs }, state) => {
-            expect(jugs.five).not.toEqual(4);
-            expect(jugs.three).toEqual(state.context.three);
-            expect(jugs.five).toEqual(state.context.five);
-          }
-        }
-      },
-      success: {
-        type: 'final',
-        meta: {
-          description: '4 gallons',
-          test: async ({ jugs }) => {
-            expect(jugs.five).toEqual(4);
-          }
-        }
-      }
-    }
-  },
-  {
-    guards: {
-      weHave4Gallons: (ctx) => ctx.five === 4
-    }
-  }
-);
-
-class Jugs {
-  public three = 0;
-  public five = 0;
-
-  public fillThree() {
-    this.three = 3;
-  }
-  public fillFive() {
-    this.five = 5;
-  }
-  public emptyThree() {
-    this.three = 0;
-  }
-  public emptyFive() {
-    this.five = 0;
-  }
-  public transferThree() {
-    const poured = Math.min(5 - this.five, this.three);
-
-    this.three = this.three - poured;
-    this.five = this.five + poured;
-  }
-  public transferFive() {
-    const poured = Math.min(3 - this.three, this.five);
-
-    this.three = this.three + poured;
-    this.five = this.five - poured;
-  }
-}
-
-const dieHardModel = createTestModel(dieHardMachine, {
-  events: {
-    POUR_3_TO_5: {
-      exec: async (_step, { jugs }) => {
-        await jugs.transferThree();
-      }
-    },
-    POUR_5_TO_3: {
-      exec: async (_step, { jugs }) => {
-        await jugs.transferFive();
-      }
-    },
-    EMPTY_3: {
-      exec: async (_step, { jugs }) => {
-        await jugs.emptyThree();
-      }
-    },
-    EMPTY_5: {
-      exec: async (_step, { jugs }) => {
-        await jugs.emptyFive();
-      }
-    },
-    FILL_3: {
-      exec: async (_step, { jugs }) => {
-        await jugs.fillThree();
-      }
-    },
-    FILL_5: {
-      exec: async (_step, { jugs }) => {
-        await jugs.fillFive();
-      }
-    }
-  }
-});
-
-describe('testing a model (shortestPathsTo)', () => {
-  dieHardModel
-    .getShortestPlansTo((state) => state.matches('success'))
-    .forEach((plan) => {
-      describe(`plan ${getDescription(plan.state)}`, () => {
-        it('should generate a single path', () => {
-          expect(plan.paths.length).toEqual(1);
-        });
-
-        plan.paths.forEach((path) => {
-          it(`path ${getDescription(path.state)}`, () => {
-            const testJugs = new Jugs();
-            return dieHardModel.testPath(path, { jugs: testJugs });
-          });
-        });
-      });
-    });
-});
-
-describe('testing a model (simplePathsTo)', () => {
-  dieHardModel
-    .getSimplePlansTo((state) => state.matches('success'))
-    .forEach((plan) => {
-      describe(`reaches state ${JSON.stringify(
-        plan.state.value
-      )} (${JSON.stringify(plan.state.context)})`, () => {
-        plan.paths.forEach((path) => {
-          it(`path ${getDescription(path.state)}`, () => {
-            const testJugs = new Jugs();
-            return dieHardModel.testPath(path, { jugs: testJugs });
-          });
-        });
-      });
-    });
-});
-
-describe('testing a model (getPlanFromEvents)', () => {
-  const plan = dieHardModel.getPlanFromEvents(
-    [
-      { type: 'FILL_5' },
-      { type: 'POUR_5_TO_3' },
-      { type: 'EMPTY_3' },
-      { type: 'POUR_5_TO_3' },
-      { type: 'FILL_5' },
-      { type: 'POUR_5_TO_3' }
-    ],
-    (state) => state.matches('success')
-  );
-
-  describe(`reaches state ${JSON.stringify(plan.state.value)} (${JSON.stringify(
-    plan.state.context
-  )})`, () => {
-    plan.paths.forEach((path) => {
-      it(`path ${getDescription(path.state)}`, () => {
-        const testJugs = new Jugs();
-        return dieHardModel.testPath(path, { jugs: testJugs });
-      });
-    });
-  });
-
-  it('should throw if the target does not match the last entered state', () => {
-    expect(() => {
-      dieHardModel.getPlanFromEvents([{ type: 'FILL_5' }], (state) =>
-        state.matches('success')
-      );
-    }).toThrow();
-  });
-});
-
-describe('.testPath(path)', () => {
-  const plans = dieHardModel.getSimplePlansTo((state) => {
-    return state.matches('success') && state.context.three === 0;
-  });
-
-  plans.forEach((plan) => {
-    describe(`reaches state ${JSON.stringify(
-      plan.state.value
-    )} (${JSON.stringify(plan.state.context)})`, () => {
-      plan.paths.forEach((path) => {
-        describe(`path ${getDescription(path.state)}`, () => {
-          it(`reaches the target state`, () => {
-            const testJugs = new Jugs();
-            return dieHardModel.testPath(path, { jugs: testJugs });
-          });
-        });
-      });
-    });
-  });
-});
-
-describe('error path trace', () => {
-  describe('should return trace for failed state', () => {
-    const machine = createMachine({
-      initial: 'first',
-      states: {
-        first: {
-          on: { NEXT: 'second' }
-        },
-        second: {
-          on: { NEXT: 'third' }
-        },
-        third: {
-          meta: {
-            test: () => {
-              throw new Error('test error');
-            }
-          }
-        }
-      }
-    });
-
-    const testModel = createTestModel(machine);
-
-    testModel
-      .getShortestPlansTo((state) => state.matches('third'))
-      .forEach((plan) => {
-        plan.paths.forEach((path) => {
-          it('should show an error path trace', async () => {
-            try {
-              await testModel.testPath(path, undefined);
-            } catch (err) {
-              expect(err.message).toEqual(
-                expect.stringContaining('test error')
-              );
-              expect(err.message).toMatchInlineSnapshot(`
-                "test error
-                Path:
-                	State: \\"first\\" |  | \\"\\"
-                	Event: {\\"type\\":\\"NEXT\\"}
-
-                	State: \\"second\\" |  | \\"\\"
-                	Event: {\\"type\\":\\"NEXT\\"}
-
-                	State: \\"third\\" |  | \\"\\""
-              `);
-              return;
-            }
-
-            throw new Error('Should have failed');
-          });
-        });
-      });
-  });
-});
-
-describe('coverage', () => {
-  it('reports state node coverage', async () => {
-    const plans = dieHardModel.getSimplePlansTo((state) => {
-      return state.matches('success') && state.context.three === 0;
-    });
-
-    for (const plan of plans) {
-      for (const path of plan.paths) {
-        const jugs = new Jugs();
-        await dieHardModel.testPath(path, { jugs });
-      }
-    }
-
-    const coverage = dieHardModel.getCoverage(stateValueCoverage());
-
-    expect(coverage.every((c) => c.status === 'covered')).toEqual(true);
-
-    expect(coverage.map((c) => [c.criterion.description, c.status]))
-      .toMatchInlineSnapshot(`
-      Array [
-        Array [
-          "Visits \\"dieHard\\"",
-          "covered",
-        ],
-        Array [
-          "Visits \\"dieHard.pending\\"",
-          "covered",
-        ],
-        Array [
-          "Visits \\"dieHard.success\\"",
-          "covered",
-        ],
-      ]
-    `);
-
-    expect(() => dieHardModel.testCoverage(stateValueCoverage())).not.toThrow();
-  });
-
-  it('tests missing state node coverage', async () => {
-    const machine = createMachine({
-      id: 'test',
-      initial: 'first',
-      states: {
-        first: {
-          on: { NEXT: 'third' }
-        },
-        secondMissing: {},
-        third: {
-          initial: 'one',
-          states: {
-            one: {
-              on: {
-                NEXT: 'two'
-              }
-            },
-            two: {},
-            threeMissing: {}
-          }
-        }
-      }
-    });
-
-    const testModel = createTestModel(machine);
-
-    const plans = testModel.getShortestPlans();
-
-    for (const plan of plans) {
-      await testModel.testPlan(plan, undefined);
-    }
-
-    expect(
-      testModel
-        .getCoverage(stateValueCoverage())
-        .filter((c) => c.status !== 'covered')
-        .map((c) => c.criterion.description)
-    ).toMatchInlineSnapshot(`
-      Array [
-        "Visits \\"test.secondMissing\\"",
-        "Visits \\"test.third.threeMissing\\"",
-      ]
-    `);
-
-    expect(() => testModel.testCoverage(stateValueCoverage()))
-      .toThrowErrorMatchingInlineSnapshot(`
-      "Coverage criteria not met:
-      	Visits \\"test.secondMissing\\"
-      	Visits \\"test.third.threeMissing\\""
-    `);
-  });
-
-  // https://github.com/statelyai/xstate/issues/729
-  it('reports full coverage when all states are covered', async () => {
-    const feedbackMachine = createMachine({
-      id: 'feedback',
-      initial: 'logon',
-      states: {
-        logon: {
-          initial: 'empty',
-          states: {
-            empty: {
-              on: {
-                ENTER_LOGON: 'filled'
-              }
-            },
-            filled: { type: 'final' }
-          },
-          on: {
-            LOGON_SUBMIT: 'ordermenu'
-          }
-        },
-        ordermenu: {
-          type: 'final'
-        }
-      }
-    });
-
-    const model = createTestModel(feedbackMachine);
-
-    for (const plan of model.getShortestPlans()) {
-      await model.testPlan(plan, null);
-    }
-
-    const coverage = model.getCoverage(stateValueCoverage());
-
-    expect(coverage).toHaveLength(5);
-
-    expect(coverage.filter((c) => c.status !== 'covered')).toHaveLength(0);
-
-    expect(() => model.testCoverage(stateValueCoverage())).not.toThrow();
-  });
-
-  it('skips filtered states (filter option)', async () => {
-    const TestBug = createMachine({
-      id: 'testbug',
-      initial: 'idle',
-      context: {
-        retries: 0
-      },
-      states: {
-        idle: {
-          on: {
-            START: 'passthrough'
-          }
-        },
-        passthrough: {
-          always: 'end'
-        },
-        end: {
-          type: 'final'
-        }
-      }
-    });
-
-    const testModel = createTestModel(TestBug);
-
-    const testPlans = testModel.getShortestPlans();
-
-    const promises: any[] = [];
-    testPlans.forEach((plan) => {
-      plan.paths.forEach(() => {
-        promises.push(testModel.testPlan(plan, undefined));
-      });
-    });
-
-    await Promise.all(promises);
-
-    expect(() => {
-      testModel.testCoverage(
-        stateValueCoverage({
-          filter: (stateNode) => {
-            return stateNode.key !== 'passthrough';
-          }
-        })
-      );
-    }).not.toThrow();
-  });
-
-  // https://github.com/statelyai/xstate/issues/981
-  it.skip('skips transient states (type: final)', async () => {
-    const machine = createMachine({
-      id: 'menu',
-      initial: 'initial',
-      states: {
-        initial: {
-          initial: 'inner1',
-
-          states: {
-            inner1: {
-              on: {
-                INNER2: 'inner2'
-              }
-            },
-
-            inner2: {
-              on: {
-                DONE: 'done'
-              }
-            },
-
-            done: {
-              type: 'final'
-            }
-          },
-
-          onDone: 'later'
-        },
-
-        later: {}
-      }
-    });
-
-    const model = createTestModel(machine);
-    const shortestPlans = model.getShortestPlans();
-
-    for (const plan of shortestPlans) {
-      await model.testPlan(plan, null);
-    }
-
-    // TODO: determine how to handle missing coverage for transient states,
-    // which arguably should not be counted towards coverage, as the app is never in
-    // a transient state for any length of time
-    model.testCoverage(stateValueCoverage());
-  });
-
-  it('tests transition coverage', async () => {
-    const model = createTestModel(
-      createMachine({
-        initial: 'a',
-        states: {
-          a: {
-            on: {
-              EVENT_ONE: 'b',
-              EVENT_TWO: 'b'
-            }
-          },
-          b: {}
-        }
-      })
-    );
-
-    await model.testPlans(model.getShortestPlans(), null);
-
-    expect(() => {
-      model.testCoverage(transitionCoverage());
-    }).toThrowErrorMatchingInlineSnapshot(`
-      "Coverage criteria not met:
-      	Transitions a on event EVENT_TWO"
-    `);
-
-    await model.testPlans(model.getSimplePlans(), null);
-
-    expect(() => {
-      model.testCoverage(transitionCoverage());
-    }).not.toThrow();
-  });
-});
+import { allStates } from '../src/coverage';
 
 describe('events', () => {
   it('should allow for representing many cases', async () => {
@@ -625,7 +70,7 @@ describe('events', () => {
       await testModel.testPlan(plan, undefined);
     }
 
-    expect(() => testModel.testCoverage(stateValueCoverage())).not.toThrow();
+    expect(() => testModel.testCoverage(allStates())).not.toThrow();
   });
 
   it('should not throw an error for unimplemented events', () => {
@@ -879,6 +324,76 @@ describe('test model options', () => {
 
     expect(testedEvents).toEqual(['NEXT', 'NEXT', 'PREV']);
   });
+
+  it('options.beforePath(...) executes before each path is tested', async () => {
+    const counts: number[] = [];
+    let count = 0;
+
+    const testModel = createTestModel(
+      createMachine({
+        initial: 'a',
+        states: {
+          a: {
+            entry: () => {
+              counts.push(count);
+            },
+            on: {
+              TO_B: 'b',
+              TO_C: 'c'
+            }
+          },
+          b: {},
+          c: {}
+        }
+      }),
+      {
+        beforePath: () => {
+          count++;
+        }
+      }
+    );
+
+    const shortestPlans = testModel.getShortestPlans();
+
+    await testModel.testPlans(shortestPlans);
+
+    expect(counts).toEqual([1, 2, 3]);
+  });
+
+  it('options.afterPath(...) executes before each path is tested', async () => {
+    const counts: number[] = [];
+    let count = 0;
+
+    const testModel = createTestModel(
+      createMachine({
+        initial: 'a',
+        states: {
+          a: {
+            entry: () => {
+              counts.push(count);
+            },
+            on: {
+              TO_B: 'b',
+              TO_C: 'c'
+            }
+          },
+          b: {},
+          c: {}
+        }
+      }),
+      {
+        afterPath: () => {
+          count++;
+        }
+      }
+    );
+
+    const shortestPlans = testModel.getShortestPlans();
+
+    await testModel.testPlans(shortestPlans);
+
+    expect(counts).toEqual([0, 1, 2]);
+  });
 });
 
 describe('invocations', () => {
@@ -906,30 +421,6 @@ describe('invocations', () => {
     });
 
     const model = createTestModel(machine, {
-      testState: (state, service) => {
-        return new Promise((res) => {
-          let actualState;
-          const t = setTimeout(() => {
-            throw new Error(
-              `expected ${state.value}, got ${actualState.value}`
-            );
-          }, 1000);
-          service.subscribe((s) => {
-            actualState = s;
-            if (s.matches(state.value)) {
-              clearTimeout(t);
-              res();
-            }
-          });
-        });
-      },
-      testTransition: (step, service) => {
-        if (step.event.type.startsWith('done.')) {
-          return;
-        }
-
-        service.send(step.event);
-      },
       events: {
         START: {
           cases: () => [
@@ -940,28 +431,48 @@ describe('invocations', () => {
       }
     });
 
-    // const plans = model.getShortestPlansTo((state) => state.matches('success'));
     const plans = model.getShortestPlans();
 
     for (const plan of plans) {
       for (const path of plan.paths) {
         const service = interpret(machine).start();
 
-        service.subscribe((state) => {
-          console.log(state.event, state.value);
-        });
+        await model.testPath(path, {
+          testState: (state) => {
+            return new Promise<void>((res) => {
+              let actualState;
+              const t = setTimeout(() => {
+                throw new Error(
+                  `expected ${state.value}, got ${actualState.value}`
+                );
+              }, 1000);
+              service.subscribe((s) => {
+                actualState = s;
+                if (s.matches(state.value)) {
+                  clearTimeout(t);
+                  res();
+                }
+              });
+            });
+          },
+          testTransition: (step) => {
+            if (step.event.type.startsWith('done.')) {
+              return;
+            }
 
-        await model.testPath(path, service);
+            service.send(step.event);
+          }
+        });
       }
     }
 
-    model.testCoverage(stateValueCoverage());
+    model.testCoverage(allStates());
   });
 });
 
 // https://github.com/statelyai/xstate/issues/1538
 it('tests transitions', async () => {
-  expect.assertions(3);
+  expect.assertions(2);
   const machine = createMachine({
     initial: 'first',
     states: {
@@ -972,15 +483,12 @@ it('tests transitions', async () => {
     }
   });
 
-  const obj = {};
-
   const model = createTestModel(machine, {
     events: {
       NEXT: {
-        exec: (step, sut) => {
+        exec: (step) => {
           expect(step).toHaveProperty('event');
           expect(step).toHaveProperty('state');
-          expect(sut).toBe(obj);
         }
       }
     }
@@ -990,7 +498,7 @@ it('tests transitions', async () => {
 
   const path = plans[0].paths[0];
 
-  await model.testPath(path, obj);
+  await model.testPath(path);
 });
 
 // https://github.com/statelyai/xstate/issues/982
